// src/app/shared/enums/printer.ts
export enum BXLTextAlignment {
  Left, Center, Right
}

export enum BXLTextAttribue {
  Default = 0,
  Bold = 2,
  Underline = 4,
  UnderlineThick = 6,
  Reverse = 8
}

export enum BXLTextSize {
  W0 = 0,
  W1 = 16,
  W2 = 32,
  W3 = 48,
  H0 = 0,
  H1 = 1,
  H2 = 2,
  H3 = 3,
}

export enum BXLPrinterStatus {
  Normal = 0,
  PaperEmpty = 1,
  CoverOpen = 2,
  PaperNearEnd = 4,
  CashDrawerHigh = 16,
  Offline = 32,
  NotOpen = 64
}